Update index.html
this is good, but when I click into the search bar the placeholder text shows. I want it to be cleared out when user clicks into the search bar. Also after they click out of the search bar AND no text has been inputted, the place holder should just be "Ask a question! Try &quot;Which sources bring in the most hires?&quot" static no animation.